Code and artifacts for evaluating and visualizing the properties of Dropout for GRU.
This commit adds all the python code and google colab notebooks that were used to train
and run the models. The project report, figures and results are also added. Actual datasets
are not attached here for obvious reasons bu their sources can be seen in the report. Check
README file for more details about reproducing results.